move fri_query_phase to fri mod
